Bundle the agent in platform-specific builds
Add a `hatch run build:all` command that invokes Hatch's build for
each of the supported platform tags. Add a build hook that downloads
the agent for the correct triple implied by that platform tag and
writes the binary to the source folder, so that Hatch can package it
in the platform-specific wheel.