Make builds a bit more usable
Add a `hatch run build:me` command that builds the package for your
current architecture. It uses some tricky logic for figuring out
what your current architecture is, which we could reuse for a
download-on-import version.

Add a `hatch run build:one <target>` command that builds the package
for an agent target of your choice.

If the `agent.py` file is updated with a new agent version, the agent
binary cache in the build hook should now take notice of that and
re-download the agent automatically.