Add Semaphore CI
Run linters and type checkers.

Test the project in Python 3.9 and above. (There are typing-related
failures in 3.8 that I did not care to investigate; Semaphore does
not support 3.7 and below)

Change the build hook to return (exiting without failure) when
no platform is given. This isn't great, but apparently Hatch does
a build when creating an environment.