Make all the config options optional fields
Don't only set the TypeDict to `total=False` to not have to specify
all the fields on initialization, but make all fields optional.
This allows us to have unset config options like booleans in the
future and recognize them by something more clear than an empty
string.
This has no influence on the interaction with these optional fields.