Add config options of the list type
Parse lists from environment variables, and format them again when set
as private env vars for the agent.

I left out request_headers, because we should check what the default for
that should be in more detail. This is more the general structure of
parsing and setting config options that the client does not directly use
(right now).