Add log_path config option
Check if the specified path is writable or not, and if not, fall back on
the system tmp dir.

It just prints the warning and doesn't log it. We don't have an internal
logger yet, see issue #34. It will also need to log it eventually.

While discussing this implementation, we decided we want to refactor
this so there's a Config class that handles all the function calls, so
that we don't need to call `validate` manually this way in the future.
See issue #36. So I'm leaving a proper validation "state" in the Config
for that issue.

[skip changeset]
Part of #16