Add internal logger
Allow us to log things to the appsignal.log file.
Add a new "appsignal" logger in Python and configure it to match our
logging standard.
https://github.com/appsignal/integration-guide/blob/e0faa6a9e02265b375a449ec6b0f234ccc68834a/build/logging.md

Use the tmp_path helper argument in pytest to create a temporary path to
store the log file in for testing.

There is no "trace" level in Python. Not sure how we can support that.

This doesn't implement the STDOUT logger. That's the next step.

Part of #34