Add request_headers config option (#52)
Allow users to specify an allow list of request headers. Configure the
OpenTelemetry instrumentation using the env var. There is no config
option on the Django or wsgi/asgi instrumentations to specify it another
way, only the env var.

There is a small difference between what you configure and what you see
in the AppSignal.com UI. The header name for the `request_headers`
config option needs to be using dashes (if the header uses dashes), but
it will show up with underscores, because that's how OpenTelemetry
exports it.

- Header name: `X-Custom-Header`
- Config: `x-custom-header` (case insensitive)
- Reported name: `x_custom_header`

I hope this won't be too confusing, let's document it properly. And we
can always change it later, but this is easiest to implement now.

[skip changeset]
Part of #16
Reapply #39 after bad merge