Lock Python versions in the project
Luismi and I were debugging our local setups and noticed a couple
problems:

- If you don't have Python 3.8 as your selected Python version, it will
  not run the test suite for it, and instead skip it.
- Hatch doesn't use the Python 3.10 and 3.11 on our machines, because it
  finds the Homebrew installed versions instead.

Add the `.python-version` file used by pyenv (the Python version manager
we use), to specify _all Python versions used in the project_. According
to this discussion [1], that allows Hatch/virtualenv to find the pyenv
installed versions.

The top one is the one you use day to day, but when running the pytest
suite through hatch, it will use the correct versions as installed with
pyenv.

You will need to delete the hatch cache located after this is merged so
you're sure it's using the right versions.

```
# macOS
rm -rf ~/Library/Application\ Support/hatch/
```

[1]: https://github.com/pypa/hatch/discussions/642