Listen and register agent exit code
When the agent is started, track if it's started correctly. This will
help debug issues and warn people when the agent can't start for
whatever reason.

I've added a global `agent_active` variable which can be accessed
through the `is_agent_active` helper function. It doesn't update if you
import the global variable directly. Useful for the diagnose tool later
on.

Listen for the exit code from the agent after starting the agent.
The Python app will wait until the agent has started and exited.

This would normally cause a TimeoutError, because the agent will need to
continue to run and `wait` waits for it to exit. However, turns out we
still daemonize the agent in this scenario (it's not marked as
standalone mode). So when it daemonizes we can assume it has started
properly. If the config is invalid according to the agent, it will exit
with an error before daemonizing.

This may add a small delay to the boot of the program, as long as the
agent needs to start and daemonize. We can possibly do this async in the
future, but I couldn't figure it out.