Load file/initial config options source load
As discussed in the integrations guide issue
https://github.com/appsignal/integration-guide/issues/73
Change the config source load order to use the "initial" config source
last.

The environment source can be referenced from the config file and
conditional behavior can be set based on it, but the given config to the
initializer will be the final used config once merged with the other
sources.

New load order:

- Default source (hardcoded config)
- System source (context dependent config)
- Environment source (environment variables)
- Initial source (any config given to `Config()` in `__appsignal__.py`)