Allow custom agent builds
An argument can be passed to `hatch run build:me` specifying the
path to the agent: `hatch run build:me path/to/agent`.

Additionally, you can use the `--keep-agent` flag to use the current
agent binary that is already present at `src/appsignal/appsignal-agent`
from a previous build: `hatch run build:me --keep-agent`.

The latter can be useful in containerised environments like our test
setups, which mount the integration repository as a volume in a
container, but not the agent repository -- meaning the location of the
agent artifact is not accessible when the integration package is
being built.