Hardcode agent version and extension loaded value
For now, hardcode the agent version and the extension loaded
result. We don't have the data for either one yet because the diagnose
command doesn't call out to the agent just yet.

Also, the agent version isn't returned by the agent, which is required
as we don't have access to the agent file in the tmp directory.