Refacotring: turn agent wrapper into a singleton
Before: agent state is a global variable
managed from a set of functions
using `global`.

After: the agent state is managed by
Agent singleton. The state is incapsulated,
and all related logic shares the same namespace.
The resulting code is shorter and easier to test.

I'd prefer to avoid a global state altogether
and let Client manage the Agent,
but that would require to rethink the tests
that currently implicitly manage
the agent state from autouse fixtures.