Print deprecation warning if beta package present
We have renamed the package to `appsignal` in PR #112. Previously, we
published the package as `appsignal-beta` because we couldn't publish
using the `appsignal` package name.

To encourage people to switch to the `appsignal` package, we will ship a
new version of the `appsignal-beta` package that has a dependency on the
new `appsignal` package and prints a deprecation warning about the beta
package if imported.

This change imports the beta package and prints the deprecation warning.
If the package is not found, it will throw an error, which we'll catch
and ignore.

Ignore the warnings from the linters about the library not being found.

Part of #111