Add host_role config option
Allow users to configure a role for a host. It's not used a lot in the
product yet, but we generate a metric for it and we can do some more
stuff with it in the future.

I chose the `host_role` name because `role` alone is a bit ambiguous and
the env var is named `APPSIGNAL_HOST_ROLE`.

Part of https://github.com/appsignal/appsignal-agent/issues/1020