Add gauge helper for reporting gauge metrics
Add convenience helper method to report gauge metrics using the
AppSignal terminology along with all the OpenTelemetry boilerplate.

This doesn't allow for setting units or metric descriptions, which
OpenTelemetry supports, because we discard this metadata.

This is using the OpenTelemetry ObservableGauge in the background, but
exposes an API to what we're used to in other integrations with
`set_gauge`.

The ObservableGauge registers a callback that it runs asynchronously in
the background every 60 seconds. 60 seconds after the app starts it
begins and when the app exits it's called one more time.

This makes the API simpler for our users so they don't have to worry
about setting up methods for callbacks and such. It's more limiting, but
people can always use the OpenTelemetry metrics module themselves to
send metrics.

More information about the OpenTelemetry metrics SDK can be found here:
https://opentelemetry-python.readthedocs.io/en/latest/sdk/metrics.html

Part of #124