Add error helpers
The `set_error` helper follows the implementation of the Node.js
integration.

For the `send_error` helper, it has been split into two: a "naked"
`send_error` function, that just sends the error without added span
context, and a `send_error_with_context` function, which must be
used with a `with` block to which the span is provided as an
argument.