Use in-memory metrics and spans in tests
Use `mocker.patch()` to turn the functions that initialize
OpenTelemetry as part of initializing the client into no-ops.

Because the meter provider and the tracer provider are globals
that can only be initialised once, use a session-scoped fixture to
initialise them.

In order to fetch them for use in a single test case, provide
function-scoped fixtures that reset their state between usages.

Implicitly fixes #88.