Validate config before starting AppSignal
When the Push API key option is not set, or blank, do not start
AppSignal as it will be able to authenticate with the server and report
no data.

No test added for the demo CLI, as it does not appear to be written.
Will be picked up as part of issue #138 hopefully.

Closes #144