Add demo CLI test and remove validate output
Add the missing test for the demo CLI. It only tests if the config is
valid or not, as testing if the OpenTelemetry span creation works is not
necessary I think for now. It will require a lot of mocking, which
doesn't tell us if the code actually still works or not.

I've also updated the config validation to not print an error
immediately on validation. This may also be printed in some scenarios we
don't want.

Preferably this is logged to the `appsignal.log` file like in the Ruby
integration, but here we don't have a logger instance yet in the Python
package. Something to improve later.

The demo CLI will log that the config is invalid, which can be debugged
with the diagnose CLI if needed. The normal client start already
logged a message when it couldn't start.

Needed for #138