Load __appsignal__.py file when found in demo CLI
When the demo CLI finds a `__appsignal__.py` configuration file, use
that for the configuration, rather than only rely on the CLI prompt.

The demo CLI will overwrite the `active` config option to always be
true when loading the file config.

One install test failed, because it expects a real file to be present
and I didn't want to add more mocking to account for that. Instead, it
now doesn't mock and uses a dummy project on the file system.

Part of #142