Fix diagnose loading config for agent diagnose
The diagnose CLI tool did not configure the agent diagnose run with the
configuration. As a result it would always fail the agent diagnose
tests.

Set the private env vars using the `Config.set_private_environ` method
and update the subprocess run of the agent with the `--private` flag to
read the right private environment variables format.

Fixes #141