Load __appsignal__.py file when found in diagnose
When the diagnose CLI finds a `__appsignal__.py` configuration file, use
that for the configuration.

Before, it relied on the default sources and environment variables, but
there was no other way to load the configuration in the diagnose report.

Update the existing changeset, as it basically covers the same material,
but for another CLI tool now too.

Closes #142