Refactor Push API key validation in install CLI (#154)
In the install CLI we had an almost identical code to the
PushApiKeyValidator module present. I've updated it to use the
PushApiKeyValidator module instead so we don't have a bunch of duplicate
code.

Change the test to expect a POST request, instead of a GET request. The
API doesn't care which HTTP request method is used, it listens to both
type of request methods.

[skip changeset] as it's an internal change with no user visible change.