Fix and improve diagnose paths report output (#159)
This change was originally prompted by the bug reported in #158.
If the `appsignal.log` file could not be found, it would exit with an
error when trying to read the log file.

While fixing this, I noticed the paths subreport didn't print the same
output as other integrations. While fixing this I refactored how the
paths subreport is generated, which also includes the fix. It was more
work to split the fix out or fix it first, than include it in the
refactor.

The fix is to early return from the method doing all the information
gathering about a path (directory/file) when it does not exist with this
if-statement in the `PathsReport._path_metadata` method:

```python
if not info["exists"]:
    return info
```

The refactor moves the paths data gathering to a new PathsReport class
for better organization. It defines which paths exists and iterates over
them to fetch the metadata and generate the paths subreport. This new
implementation reduces duplication in definitions of paths and path
output, and removes a second method that also read the file contents.
Instead the CLI will only print the last 10 lines of file contents using
the full file contents value (which is the last 2MB of the file).

Fixes #158