Bump package to version 1.0 (#155)
Nothing in this release requires a major version bump, but as discussed,
we are ready to release version 1.0 to communicate it's ready for use.

It's not feature complete: we don't have Python logging integration for
example, but those features will be added a later time.

I created a small changeset that kind of summarizes what the package
contains now, but this is not meant to replace the blog post
announcement and we should update the changelog later to link to that
instead. This is there now in case we release before a blog post is out.

The changeset file name starts with an underscore to hopefully put it on
the top upon release.