Report agent architecture and platform in diagnose
We ship a separate package for every architecture and platform. The
agent isn't installed upon package download/install, like in other
integrations.

By adding the agent's architecture and platform, we know which package
version was installed for which architecture and platform. Ship a
`_appsignal_platform` file in the distributed package, which is then
read by the diagnose CLI. I don't want the diagnose CLI to detect the
host here, I want to know which release artifact (as pushed to pypi) was
installed.