Report running_in_container in diagnose report
I've updated the agent to report the `running_in_container` field in the
agent's diagnose report's `host` sub report.

In the Python package I've copied it to the host sub report on the main
diagnose report as well. That way the diagnose viewer has to less about
where this field is originates from for every integration. It will
report this field for the agent report, as for all other integrations,
but that won't get in the way. We can even hide it in the viewer.

Part of https://github.com/appsignal/appsignal-python/issues/172