Fix demo CLIs samples metadata
Remove unused `http.method` attributes. They would appear in the
"Attributes" box in the UI that only shows up for root span attributes
we don't know what to do with.

Set the category for the performance sample so the event timeline
doesn't show an "unknown" event.

Removing the `otel.instrumentation_library.name` attribute because it's
not needed. It will use the root span's "GET /demo" name by default.