Fix missing platform file error
The platform file should never be missing. But if it somehow is,
then don't crash the diagnose report because of it.

Fix a typo in the name of the function that reads from the platform
file.

If the platform file is missing from our agent build cache, make
sure that we do not consider that to be a correct build. Downloading
the agent again is overkill, but it does the trick.

Ensure the platform file is actually added by declaring it as a
build artifact.