Remove sudo from CI install command for hatch
Not sure why we were installing it with `sudo` but that broke the
`sem-version python 3.11` setting, and ran the linter under Python 3.8
(the system default) instead.

Remove `sudo` from the command so the linter runs in the Python version
we configure.

Because some jobs override the Python version set in the global
prologue, move it to each individual job to avoid confusion.