Do not use `@cached_property` for prompts
Using `@cached_property` to define these prompting functions (which
are, fundamentally, not really properties at all) makes it a lot
harder to reason about what happens when you invoke them. Furthermore,
as far as I'm aware, there's no documented way to clear the cache for
a cached property.

This change is, in the wise words of Mickey Mouse, "a surprise tool
that will help us later".