Test asking again for API key
Make it so that the `mock_input` context manager can indicate the
expectation of an unanswered prompt by using `None` instead of an
answer.

When this happens, the expected `StopIteration` that indicates that
a prompt was asked for which no answer was given is not raised as an
error.

Delete several duplicates of `mock_input` and move it to a single
`utils.py` file.