Set metrics export interval to ten seconds
Note that this is the export interval for metrics to be sent from
the OpenTelemetry application to the agent, not from the agent to
AppSignal's servers.

The default interval is set to sixty seconds -- when combined with
the agent's transmission interval, this introduces a delay of up to
two minutes in metrics being sent to AppSignal.

In addition, the histogram-to-distribution conversion that is
performed in the agent benefits from smaller batches of data for
greater accuracy.