Test `start` and `stop` methods
Got in a bit of a pickle to make the thread initialization testable.
When `client.start()` was called, a thread was spawned, so I had to
implement a way to stop these threads from executing during other
tests (hence `client.stop()`)

All other tests continue to call `_run_probes` directly because the
resulting behaviour is easier to reason about.