Implement heartbeats for Python
Implement the heartbeats functionality for Python, similarly to how
it was implemented for Ruby, Elixir and Node.js.

Due to limitations on how `with` blocks are implemented, which make
it impossible for a function to behave in different ways when called
"by itself" and when called by a `with` block, it is possible to call
it in the following ways:

- `heartbeat(name)`, which sends a finish heartbeat event
- `heartbeat(name, function)`, which sends:
  - a start heartbeat event
  - a finish heartbeat event, if the function does not raise
- as a `with Heartbeat(name):` block (note: `H` not `h`!) which sends:
  - a start heartbeat event
  - a finish heartbeat event, if the `with` block does not raise

The latter is intentionally un-documented for now, to avoid potential
confusion between `heartbeat` and `Heartbeat`.

Add a `LOGGING_ENDPOINT` configuration variable to configure the
logging endpoint, which is also the endpoint to which heartbeats are
sent. This is a bit bizarre, as Python does not support the logging
feature yet, but it is consistent with our other integrations.