Move internal logger to separate module
The logger was kind of "bolted on" to the client class -- when other
parts of the code wanted to use it, they needed to fetch the logger
with the same name separately and deal with it as a variable.

Move this to a separate `internal_logger` module which is initialised
on-demand from the existing initialised (not necessarily started!)
global client.