Add `collector_endpoint` config option
Add a configuration option that allows for a collector endpoint to
be provided. When given, AppSignal for Python will not use the agent
binary that is bundled with it, and instead send data to the given
collector endpoint.

Implement this by defining a `Binary` base class, and swapping the
AppSignal's client `Agent` with a `NoopBinary` when the collector
endpoint configuration option is set. This paves the way for defining
a `Collector` class later on.