Configure OpenTelemetry for collector
Configure the OpenTelemetry resource attributes to provide the
required attributes for the collector to process the data.

Add a new `service_name` configuration option to provide the service
name that the collector expects.

Since `revision` is optional, set a placeholder value when it is not
present. Do the same for `service_name`.