Instrument logging when using the collector
When the collector is used, configure the OpenTelemetry logging SDK.

Add an instrumentation that adds OpenTelemetry's `LoggingHandler`
to the root logger in Python's built-in `logging` module. This will
automatically send logs to AppSignal from any loggers that are not
explicitly configured not to propagate to the root logger.

Configure AppSignal's internal logger not to propagate to the root
logger.