Simplify default instrumentation names
The convention of using package names made sense as long as what we
were enabling or disabling mapped cleanly to them. In the `logging`
case, we're enabling instrumentation for the `logging` package, but
this instrumentation lives in the `opentelemetry.sdk._logs` package.
Using that package name, however, is confusing, as someone familiar
with the SDK would think that we're disabling the logs section of the
OpenTelemetry SDK entirely.

Change the default instrumentation names to be the names of the
libraries that they instrument, rather than the module paths of the
modules in which the instrumentations are added. This also makes them
shorter, which I'm sure is nice if you need to specify a couple of
them.

For backwards compatibility, support the name being specified with
an `opentelemetry.instrumentation.` prefix, which is ignored when
matching the names.