Add collector attributes config options
Implement the following configuration options in the integration,
mapping to the collector-supported, `appsignal.config.`-prefixed
resource attribute of the same name:

- `filter_attributes`
- `filter_function_parameters`
- `filter_request_query_parameters`
- `filter_request_payload`
- `response_headers`
- `send_function_parameters`
- `send_request_query_parameters`
- `send_request_payload`

Additionally, also map the following existing configuration options
in the integrations to resource attributes in the same manner:

- `filter_session_data` (mapped to `filter_request_session_data`)
- `ignore_actions`
- `ignore_errors`
- `ignore_namespaces`
- `request_headers`
- `send_session_data` (mapped to `send_request_session_data`)

When validating the configuration, if the configuration would trigger
the use of the collector, emit a warning about any configuration
options set to non-default values which have no effect when using
the collector. Do the same when the agent would be used for
configuration options that would have no effect when using the agent.