Fix custom agent build instructions
Currently, README.md suggests that running `hatch run build:me
--keep-agent` will keep the agent binary from the `src/appsignal`
directory intact. But as a part of the build environment setup,
`hatch` installs project in development mode, which runs the
`src/scripts/build_hook.py` script. This results in the script
discarding the local binary, despite the `--keep-agent` flag.

This change adds `_APPSIGNAL_BUILD_AGENT_PATH` env variable to the build
script, so that it is picked up by the project installation and
build steps.

[skip ci]