Add `requests` dependency to `hatch-build` env
After changes in Hatch 1.16, it seems that the build hook script is
sometimes ran within the `hatch-build` environment, rather than in
the `build` environment it is declared for. As such, the `requests`
dependency for the build hook must be declared for this environment,
as well as for the script itself.